refactor Query resolver
